+++ src/tools/formatters.ts
@@ -22,7 +22,8 @@ const ranks = [
 ];
 
 export async function scoreList(
-    scores: apitypes.Score[],
+    scores: apitypes.ScoreLegacy[] | apitypes.Score[],
+    scoretype: 'legacy' | 'current',
     sort: 'pp' | 'score' | 'recent' | 'acc' | 'combo' | 'miss' | 'rank',
     filter: {
         mapper: string,
@@ -48,7 +49,10 @@ export async function scoreList(
     showMap?: boolean,
     overrideMap?: apitypes.Beatmap
 ): Promise<formatterInfo> {
-    const newScores = filterScores(scores, sort, filter, reverse, overrideMap);
+    const newScores =
+        scoretype == 'legacy' ?
+            filterScoresLegacy(scores as apitypes.ScoreLegacy[], sort, filter, reverse, overrideMap) :
+            filterScores(scores as apitypes.Score[], sort, filter, reverse, overrideMap);
     if (newScores.length == 0) {
         return {
             text: 'No scores were found (check the filter options)',
@@ -65,46 +69,66 @@ export async function scoreList(
     let text = '';
 
     for (let i = 0; i < max && i < newScores.length - offset; i++) {
-        const score = newScores[i + offset];
+        let score = newScores[i + offset];
         if (!score) break;
+        let convertedScore = score as apitypes.ScoreLegacy;
+        if (scoretype == 'current') {
+            convertedScore = CurrentToLegacyScore(score as apitypes.Score);
+        }
         const perf = await helper.tools.performance.calcScore({
-            mapid: (overrideMap ?? score.beatmap).id,
-            mode: score.mode,
-            mods: score.mods.join(''),
+            mapid: (overrideMap ?? convertedScore.beatmap).id,
+            mode: convertedScore.mode,
+            mods: convertedScore.mods.join(''),
             accuracy: score.accuracy,
-            hit300: score.statistics.count_300,
-            hit100: score.statistics.count_100,
-            hit50: score.statistics.count_50,
-            miss: score.statistics.count_miss,
-            hitkatu: score.statistics.count_katu,
-            maxcombo: score.max_combo,
-            mapLastUpdated: new Date((overrideMap ?? score.beatmap).last_updated),
+            hit300: convertedScore.statistics.count_300,
+            hit100: convertedScore.statistics.count_100,
+            hit50: convertedScore.statistics.count_50,
+            miss: convertedScore.statistics.count_miss,
+            hitkatu: convertedScore.statistics.count_katu,
+            maxcombo: convertedScore.max_combo,
+            mapLastUpdated: new Date((overrideMap ?? convertedScore.beatmap).last_updated),
         });
         const fc = await helper.tools.performance.calcFullCombo({
             mapid: (overrideMap ?? score.beatmap).id,
-            mode: score.mode,
-            mods: score.mods.join(''),
-            accuracy: score.accuracy,
-            hit300: score.statistics.count_300,
-            hit100: score.statistics.count_100,
-            hit50: score.statistics.count_50,
-            hitkatu: score.statistics.count_katu,
-            maxcombo: score.max_combo,
+            mode: convertedScore.mode,
+            mods: convertedScore.mods.join(''),
+            accuracy: convertedScore.accuracy,
+            hit300: convertedScore.statistics.count_300,
+            hit100: convertedScore.statistics.count_100,
+            hit50: convertedScore.statistics.count_50,
+            hitkatu: convertedScore.statistics.count_katu,
+            maxcombo: convertedScore.max_combo,
             mapLastUpdated: new Date((overrideMap ?? score.beatmap).last_updated),
         });
         let info = `**#${(showOriginalIndex ? score.originalIndex : i) + 1}`;
         if (showMap != false) {
-            info += `・[${score.beatmapset.title} [${(overrideMap ?? score.beatmap).version}]](https://osu.ppy.sh/scores/${score.mode}/${score.id})`;
+            if (scoretype == 'legacy') {
+                score = score as indexedScore<apitypes.ScoreLegacy>;
+                info += `・[${score.beatmapset.title} [${(overrideMap ?? score.beatmap).version}]](https://osu.ppy.sh/scores/${score.mode}/${score.id})`;
+            } else {
+                score = score as indexedScore<apitypes.Score>;
+                info += `・[${score.beatmapset.title} [${(overrideMap ?? score.beatmap).version}]](https://osu.ppy.sh/scores/${score.id})`;
+            }
         }
         if (showUsername) {
             info += `・[${score.user.username}](https://osu.ppy.sh/u/${score.user_id})`;
         }
-        info +=
-            `**
-\`${helper.tools.calculate.numberShorthand(score.score)}\` |${score.mods.length > 0 ? ' **' + score.mods.join('') + '** |' : ''} ${dateToDiscordFormat(new Date(score.created_at))}
-\`${hitList(score.mode, score.statistics)}\` | ${score.max_combo} | ${(score.accuracy * 100).toFixed(2)} | ${helper.vars.emojis.grades[score.rank]}
-${(score?.pp ?? perf.pp).toFixed(2)}pp`;
-        if (!score.perfect) {
+        if (scoretype == 'legacy') {
+            score = score as indexedScore<apitypes.ScoreLegacy>;
+            info +=
+                `**
+    \`${helper.tools.calculate.numberShorthand(score.score)}\` |${score.mods.length > 0 ? ' **' + score.mods.join('') + '** |' : ''} ${dateToDiscordFormat(new Date(score.created_at))}
+    \`${hitList(score.mode, score.statistics)}\` | ${score.max_combo} | ${(score.accuracy * 100).toFixed(2)} | ${helper.vars.emojis.grades[score.rank]}
+    ${(score?.pp ?? perf.pp).toFixed(2)}pp`;
+        } else {
+            const tempScore = score as indexedScore<apitypes.Score>;
+            info +=
+                `**
+    \`${helper.tools.calculate.numberShorthand(tempScore.total_score)}\` |${tempScore.mods.length > 0 ? ' **' + tempScore.mods.map(x => x.acronym).join('') + '** |' : ''} ${dateToDiscordFormat(new Date(tempScore.ended_at))}
+    \`${hitList(convertedScore.mode, convertedScore.statistics)}\` | ${score.max_combo} | ${(score.accuracy * 100).toFixed(2)} | ${helper.vars.emojis.grades[score.rank]}
+    ${(score?.pp ?? perf.pp).toFixed(2)}pp`;
+        }
+        if (!convertedScore?.perfect) {
             info += '(' + fc.pp.toFixed(2) + 'pp if FC)';
         }
         info += '\n\n';
@@ -128,12 +152,12 @@ ${(score?.pp ?? perf.pp).toFixed(2)}pp`;
     };
 }
 
-type indexedScore = apitypes.Score & {
+type indexedScore<T> = T & {
     originalIndex: number,
 };
 
-export function filterScores(
-    scores: apitypes.Score[],
+export function filterScoresLegacy(
+    scores: apitypes.ScoreLegacy[],
     sort: 'pp' | 'score' | 'recent' | 'acc' | 'combo' | 'miss' | 'rank',
     filter: {
         mapper: string,
@@ -153,8 +177,8 @@ export function filterScores(
     },
     reverse: boolean,
     overrideMap?: apitypes.Beatmap
-) {
-    let newScores: indexedScore[] = [];
+): indexedScore<apitypes.ScoreLegacy>[] {
+    let newScores = [] as indexedScore<apitypes.ScoreLegacy>[];
     for (let i = 0; i < scores.length; i++) {
         const newScore = { ...scores[i], ...{ originalIndex: i } };
         newScores.push(newScore);
@@ -285,6 +309,169 @@ export function filterScores(
             break;
         case 'miss':
             newScores.sort((a, b) => a.statistics.count_miss - b.statistics.count_miss);
+            break;
+        case 'rank':
+            newScores.sort((a, b) => ranks.indexOf(a.rank) - ranks.indexOf(b.rank));
+            break;
+    }
+    if (reverse) newScores.reverse();
+
+    return newScores;
+}
+
+export function filterScores(
+    scores: apitypes.Score[],
+    sort: 'pp' | 'score' | 'recent' | 'acc' | 'combo' | 'miss' | 'rank',
+    filter: {
+        mapper: string,
+        title: string,
+        artist: string,
+        version: string,
+        modsInclude: string,
+        modsExact: string,
+        modsExclude: string,
+        rank: string,
+        pp: string,
+        score: string,
+        acc: string,
+        combo: string,
+        miss: string,
+        bpm: string,
+    },
+    reverse: boolean,
+    overrideMap?: apitypes.Beatmap
+): indexedScore<apitypes.Score>[] {
+    let newScores = [] as indexedScore<apitypes.Score>[];
+    for (let i = 0; i < scores.length; i++) {
+        const newScore = { ...scores[i], ...{ originalIndex: i } };
+        newScores.push(newScore);
+    }
+    if (filter.mapper) {
+        newScores = newScores.filter(score =>
+            matchesString(score.beatmapset.user.username, filter.mapper) || matchesString(score.beatmapset.user_id + '', filter.mapper) || matchesString((overrideMap ?? score.beatmap).user_id + '', filter.mapper));
+    }
+    if (filter.title) {
+        newScores = newScores.filter(score =>
+            matchesString(score.beatmapset.title, filter.title) || matchesString(score.beatmapset.title_unicode, filter.title));
+    }
+    if (filter.artist) {
+        newScores = newScores.filter(score =>
+            matchesString(score.beatmapset.artist, filter.artist) || matchesString(score.beatmapset.artist_unicode, filter.artist));
+    }
+    if (filter.version) {
+        newScores = newScores.filter(score =>
+            matchesString((overrideMap ?? score.beatmap).version, filter.version));
+    }
+    if (filter.pp) {
+        const tempArg = argRange(filter.pp, true);
+        newScores = newScores.filter(score =>
+            !isNaN(tempArg.max) && !isNaN(tempArg.min) && !isNaN(tempArg.exact) ?
+                tempArg.max ?
+                    score.pp <= tempArg.max :
+                    tempArg.min ?
+                        score.pp >= tempArg.min :
+                        score.pp == tempArg.exact : true);
+    }
+    if (filter.score) {
+        const tempArg = argRange(filter.score, true);
+        newScores = newScores.filter(score =>
+            !isNaN(tempArg.max) && !isNaN(tempArg.min) && !isNaN(tempArg.exact) ?
+                tempArg.max ?
+                    score.total_score <= tempArg.max :
+                    tempArg.min ?
+                        score.total_score >= tempArg.min :
+                        score.total_score == tempArg.exact : true);
+    }
+    if (filter.acc) {
+        const tempArg = argRange(filter.acc, true);
+        newScores = newScores.filter(score =>
+            !isNaN(tempArg.max) && !isNaN(tempArg.min) && !isNaN(tempArg.exact) ?
+                tempArg.max ?
+                    score.accuracy * 100 <= tempArg.max :
+                    tempArg.min ?
+                        score.accuracy * 100 >= tempArg.min :
+                        score.accuracy * 100 == tempArg.exact : true);
+    }
+    if (filter.combo) {
+        const tempArg = argRange(filter.combo, true);
+        newScores = newScores.filter(score =>
+            !isNaN(tempArg.max) && !isNaN(tempArg.min) && !isNaN(tempArg.exact) ?
+                tempArg.max ?
+                    score.max_combo <= tempArg.max :
+                    tempArg.min ?
+                        score.max_combo >= tempArg.min :
+                        score.max_combo == tempArg.exact : true);
+    }
+    if (filter.miss) {
+        const tempArg = argRange(filter.miss, true);
+        newScores = newScores.filter(score =>
+            !isNaN(tempArg.max) && !isNaN(tempArg.min) && !isNaN(tempArg.exact) ?
+                tempArg.max ?
+                    (score.statistics.miss ?? 0) <= tempArg.max :
+                    tempArg.min ?
+                        (score.statistics.miss ?? 0) >= tempArg.min :
+                        (score.statistics.miss ?? 0) == tempArg.exact : true);
+    }
+    if (filter.bpm) {
+        const tempArg = argRange(filter.bpm, true);
+        newScores = newScores.filter(score =>
+            !isNaN(tempArg.max) && !isNaN(tempArg.min) && !isNaN(tempArg.exact) ?
+                tempArg.max ?
+                    (overrideMap ?? score.beatmap).bpm <= tempArg.max :
+                    tempArg.min ?
+                        (overrideMap ?? score.beatmap).bpm >= tempArg.min :
+                        (overrideMap ?? score.beatmap).bpm == tempArg.exact : true);
+    }
+    if (filter.modsInclude?.includes('NM')) {
+        filter.modsExact = filter.modsInclude.replace('NM', '');
+        filter.modsInclude = null;
+    }
+    if (filter.modsInclude) {
+        newScores = newScores.filter(score => {
+            let x: boolean = true;
+            score.mods.forEach(mod => {
+                if (!osumodcalc.modHandler(filter.modsInclude, osumodcalc.ModeIntToName(score.ruleset_id)).includes(mod.acronym as osumodcalc.Mods)) {
+                    x = false;
+                }
+            });
+            return x;
+        });
+    }
+    if (filter.modsExact && !filter.modsInclude) {
+        if (['NM', 'NONE', 'NO', 'NOMOD'].some(mod => mod == filter.modsExact.toUpperCase())) {
+            newScores = newScores.filter(score => score.mods.length == 0);
+        } else {
+            newScores = newScores.filter(score => score.mods.map(x => x.acronym).join('') == osumodcalc.modHandler(filter.modsExact, osumodcalc.ModeIntToName(score.ruleset_id)).join(''));
+        }
+    } else if (filter.modsExclude) {
+        newScores = newScores.filter(score => {
+            let x: boolean = true;
+            score.mods.forEach(mod => {
+                if (osumodcalc.modHandler(filter.modsInclude, osumodcalc.ModeIntToName(score.ruleset_id)).includes(mod.acronym as osumodcalc.Mods)) {
+                    x = false;
+                }
+            });
+            return x;
+        });
+    }
+    switch (sort) {
+        case 'pp':
+            newScores.sort((a, b) => b.pp - a.pp);
+            break;
+        case 'score':
+            newScores.sort((a, b) => b.total_score - a.total_score);
+            break;
+        case 'recent':
+            newScores.sort((a, b) => (new Date(b.ended_at ?? b.started_at)).getTime() - (new Date(a.ended_at ?? a.started_at)).getTime());
+            break;
+        case 'acc':
+            newScores.sort((a, b) => b.accuracy - a.accuracy);
+            break;
+        case 'combo':
+            newScores.sort((a, b) => b.max_combo - a.max_combo);
+            break;
+        case 'miss':
+            newScores.sort((a, b) => (a.statistics.miss ?? 0) - (b.statistics.miss ?? 0));
             break;
         case 'rank':
             newScores.sort((a, b) => ranks.indexOf(a.rank) - ranks.indexOf(b.rank));
@@ -876,4 +1063,49 @@ export function removeURLparams(url: string) {
         return url.split('?')[0];
     }
     return url;
+}
+
+/**
+ * converts a lazer score to legacy score (for compatibility reasons)
+ */
+export function CurrentToLegacyScore(score: apitypes.Score): apitypes.ScoreLegacy {
+    return {
+        accuracy: score.accuracy,
+        beatmap: score.beatmap,
+        beatmapset: score.beatmapset,
+        best_id: score.best_id,
+        created_at: score.ended_at ?? score.started_at,
+        id: score.legacy_score_id,
+        match: null,
+        max_combo: score.max_combo,
+        mode_int: score.ruleset_id,
+        mode: osumodcalc.ModeIntToName(score.ruleset_id),
+        mods: score.mods.map(x => x.acronym),
+        passed: score.passed,
+        perfect: score.is_perfect_combo,
+        pp: score.pp,
+        preserve: score.preserve,
+        processed: score.processed,
+        rank_country: score.rank_country,
+        rank_global: score.rank_global,
+        rank: score.rank,
+        replay: score.has_replay,
+        room_id: score.room_id,
+        ruleset_id: score.ruleset_id,
+        score: score.legacy_total_score,
+        scores_around: score.scores_around,
+        statistics: {
+            count_100: score.statistics?.ok ?? 0,
+            count_300: score.statistics.great,
+            count_50: score.statistics?.meh ?? 0,
+            count_geki: 0,
+            count_katu: 0,
+            count_miss: score.statistics?.miss ?? 0,
+        },
+        type: score.type,
+        user_id: score.user_id,
+        current_user_attributes: score.current_user_attributes,
+        user: score.user,
+        weight: score.weight
+    };
 }